feat: migrate-Feature-Flag-APM-method-to-Core
