+++ android/native.gradle
@@ -1,5 +1,5 @@
 project.ext.instabug = [
-    version: '13.2.0'
+    version: '13.2.0.6091772-SNAPSHOT'
 ]
 
 dependencies {
