+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugAPMModule.java
@@ -7,33 +7,20 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.facebook.react.bridge.Callback;
-import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReactContext;
-import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableMap;
-import com.facebook.react.bridge.WritableMap;
 import com.instabug.apm.APM;
-import com.instabug.apm.configuration.cp.APMFeature;
-import com.instabug.apm.configuration.cp.APMFeaturesAvailability;
 import com.instabug.apm.model.ExecutionTrace;
 import com.instabug.apm.networking.APMNetworkLogger;
 import com.instabug.apm.networkinterception.cp.APMCPNetworkLog;
 import com.instabug.reactlibrary.utils.EventEmitterModule;
 import com.instabug.reactlibrary.utils.MainThreadHandler;
-import com.instabug.apm.InternalAPM;
-import com.instabug.apm.configuration.cp.FeaturesChangeListener;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
 import java.util.HashMap;
-import java.util.Properties;
 
 import javax.annotation.Nonnull;
 
@@ -323,104 +310,6 @@ public class RNInstabugAPMModule extends EventEmitterModule {
                     e.printStackTrace();
                 }
             }
-        });
-    }
-    /**
-     * Register a listener for W3C flags value change
-     */
-    @ReactMethod
-    public void registerW3CFlagsChangeListener(final Callback handler){
-
-        MainThreadHandler.runOnMainThread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    InternalAPM._registerCPFeaturesChangeListener(new FeaturesChangeListener() {
-                        @Override
-                        public void invoke(@NonNull APMFeaturesAvailability apmFeaturesAvailability) {
-                            WritableMap params = Arguments.createMap();
-                            params.putBoolean("isW3ExternalTraceIDEnabled", apmFeaturesAvailability.isW3CExternalTraceIdAvailable());
-                            params.putBoolean("isW3ExternalGeneratedHeaderEnabled", apmFeaturesAvailability.getShouldAttachGeneratedHeader());
-                            params.putBoolean("isW3CaughtHeaderEnabled", apmFeaturesAvailability.getShouldAttachCapturedHeader());
-
-                            sendEvent(Constants.IBGAPM_ON_NEW_W3C_FLAGS_UPDATE_RECEIVED_CALLBACK, params);
-                        }
-                    });
-                }
-                catch (Exception e) {
-                    e.printStackTrace();
-                }
-
-            }
-
-        });
-    }
-
-
-    /**
-     *  Get first time Value of W3ExternalTraceID flag
-     */
-    @ReactMethod
-    public void isW3ExternalTraceIDEnabled(Promise promise){
-
-        MainThreadHandler.runOnMainThread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    promise.resolve(InternalAPM._isFeatureEnabledCP(APMFeature.W3C_EXTERNAL_TRACE_ID, " "));
-                }
-                catch (Exception e) {
-                    e.printStackTrace();
-                    promise.resolve(null);
-                }
-
-            }
-
-        });
-    }
-
-
-    /**
-     *  Get first time Value of W3ExternalGeneratedHeader flag
-     */
-    @ReactMethod
-    public void isW3ExternalGeneratedHeaderEnabled(Promise promise){
-
-        MainThreadHandler.runOnMainThread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    promise.resolve(InternalAPM._isFeatureEnabledCP(APMFeature.W3C_GENERATED_HEADER_ATTACHING, " "));
-                }
-                catch (Exception e) {
-                    e.printStackTrace();
-                    promise.resolve(null);
-                }
-
-            }
-
-        });
-    }
-
-    /**
-     *  Get first time Value of W3CaughtHeader flag
-     */
-    @ReactMethod
-    public void isW3CaughtHeaderEnabled(Promise promise){
-
-        MainThreadHandler.runOnMainThread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    promise.resolve(InternalAPM._isFeatureEnabledCP(APMFeature.W3C_CAPTURED_HEADER_ATTACHING,""));
-                }
-                catch (Exception e) {
-                    e.printStackTrace();
-                    promise.resolve(null);
-                }
-
-            }
-
         });
     }
