+++ src/native/NativeAPM.ts
@@ -48,26 +48,8 @@ export interface ApmNativeModule extends NativeModule {
   startUITrace(name: string): void;
   endUITrace(): void;
   ibgSleep(): void;
-
-  // W3C Feature Flags
-  isW3ExternalTraceIDEnabled(): Promise<boolean>;
-  isW3ExternalGeneratedHeaderEnabled(): Promise<boolean>;
-  isW3CaughtHeaderEnabled(): Promise<boolean>;
-
-  // W3C Feature Flags Listener for Android
-  registerW3CFlagsChangeListener(
-    handler: (payload: {
-      isW3ExternalTraceIDEnabled: boolean;
-      isW3ExternalGeneratedHeaderEnabled: boolean;
-      isW3CaughtHeaderEnabled: boolean;
-    }) => void,
-  ): void;
 }
 
 export const NativeAPM = NativeModules.IBGAPM;
-
-export enum NativeEvents {
-  ON_W3C_FLAGS_CHANE = 'IBGAPMOnNewW3CFlagsUpdateReceivedCallback',
-}
 
 export const emitter = new NativeEventEmitter(NativeAPM);
